+++ R/getElevation.R
@@ -4,9 +4,10 @@
 #' @author Josh Roberti \email{jaroberti87@@gmail.com} \cr
 
 #' @description Return metadata of environmental monitoring stations that have a specific elevation.
-#' @param elevThresh (numeric) defines elevation range to filter metadata.  Units are in meters (m).  If \code{elevThresh} is a single value, the function will return sites within the database that have elevations less than or equal to \code{elevThresh}. Alternatively, if \code{elevThresh} is a numeric vector of length = 2, the function will assign the first component as a midpoint elevation, and the second component as a threshold (range), e.g., \code{elevThresh}=c(100,50) will return sites that have elevations within 100 +/- 50 (m) Above Sea Level.
+#' @param elevMin (numeric) defines the minimum elevation to filter metadata.  Units are in meters (m) Above Sea Level.
+#' @param elevMax (numeric) defines the maximum elevation to filter metadata.  Units are in meters (m) Above Sea Level.
 #'
-#'@param ... auto-populates when called from \code{siteFinder()} wrapper
+#' @param ... auto-populates when called from \code{siteFinder()} wrapper
 #'
 #' @return A list comprising metadata of environmental monitoring stations that have elevations conforming to the criteria specified in \code{elevThresh}\cr
 
@@ -14,10 +15,10 @@
 
 #' @examples
 #' \dontrun{
-#' #return metadata of sites that have elevations between 1500 +/- 500 (m) Above Sea Level
-#'   getElevation(elevThresh=c(1500,500))
-#' #return metadata of sites that have elevations less than or equal to 3500 (m) Above Sea Level
-#'   getElevation(elevThresh=3500)}
+#' #return metadata of sites that have elevations above 1500 (m) Above Sea Level
+#'   getElevation(elevMin=1500)
+#' #return metadata of sites that have elevations less than 35 (m) Above Sea Level
+#'   getElevation(elevMax=35)}
 
 #' @seealso
 #' \link[metScanR]{siteFinder}
@@ -33,36 +34,54 @@
 #       function created as standalone or for use in siteFinder()
 #   Josh Roberti (2017-05-21)
 #       Removing NULL initializations, replacing with missing() internally
+#   Robert Lee (2018-02-09)
+#       Changing function to take elevMax and elevMin instead of elevThresh
 ##############################################################################################
 
-getElevation<-function(elevThresh,...){
+getElevation<-function(elevMin, elevMax,...){
     metadata<-c(...)
-    #if using external of wrapper:
     if(is.null(metadata)){
         metadata<-metScanR_DB
     }
-    if(!missing(elevThresh)){
+    # inArgs<-as.list(sys.call())
+    # #Deprecated parameters elevThresh replaced with elevMin and elevMax
+    # # #Will officially remove these terms in a future release
+    # if(!is.null(inArgs$elevThresh)){
+    #     .Deprecated(new="elevMin and/or elevMax",msg="The 'elevThresh' parameter is depracated.  The user is asked to use 'minElev' and/or 'maxElev' as replacements. See ?getElevation for more information.",old=c(inArgs$ele))
+    #     #assign NEON.site, Lat, and Lon to updated parameters so code still works:
+    #
+    #     if(length(inArgs$elevThresh)==1){
+    #         elevMin<-0
+    #         elevMax<-inArgs$elevThresh
+    #     }
+    #     if(length(inArgs$elevThresh)==2){
+    #         elevMin<-elevThresh[1]-inArgs$elevThresh[2]
+    #         elevMax<-elevThresh[1]+inArgs$elevThresh[2]
+    #     }
+    #     if(length(inArgs$elevThresh)>2){
+    #         #if length(elevThresh) >2:
+    #         stop("Incorrect number of inputs")
+    #     }
+    # }
+    if(!missing(elevMin)&missing(elevMax)){ #min elevation up
         elevation<-lapply(lapply(metadata,
                                  "[[","location"),"[[", "elev")
-        if(length(elevThresh)==1){
-            #find elevations that meet criterion
-            elevationMatch<-which(elevation<=elevThresh)
-            #return filtered metadata
-            metadata<-metadata[match(names(elevationMatch),names(metadata))]
-        }
-        if(length(elevThresh)==2){
-            #create lower and upper bounds
-            lowerBound<-elevThresh[1]-elevThresh[2]
-            upperBound<-elevThresh[1]+elevThresh[2]
-            #find elevations that meet criteria
-            elevationMatch<-which(elevation>=lowerBound & elevation<=upperBound)
-            #return filtered metadata
-            metadata<-metadata[match(names(elevationMatch),names(metadata))]
-        }
-        if(length(elevThresh)>2){
-            #if length(elevThresh) >2:
-            stop("Incorrect number of inputs")
-        }
+        elevationMatch<-which(elevation>=elevMin)
+        #return filtered metadata
+        metadata<-metadata[match(names(elevationMatch),names(metadata))]
+    }
+    if(missing(elevMin)&!missing(elevMax)){ #max elevation down
+        elevation<-lapply(lapply(metadata,
+                                 "[[","location"),"[[", "elev")
+         elevationMatch<-which(elevation<=elevMax)
+        #return filtered metadata
+        metadata<-metadata[match(names(elevationMatch),names(metadata))]
+    }
+    if(!missing(elevMin)&!missing(elevMax)){
+        elevation<-lapply(lapply(metadata,
+                                 "[[","location"),"[[", "elev")
+        elevationMatch<-which(elevation>=elevMin & elevation<=elevMax)
+        metadata<-metadata[match(names(elevationMatch),names(metadata))]
     }
     else{
         #if getElevation is NULL: